Move refine-rule resolution from xccdf_policy.c to xccdf_policy_resolve.c
